New script and publish wb
New script and publish wb
